Split diagnostics helper modules
