+++ apps/server/tests/use_cases/diagnostics/test_run_analysis.py
@@ -8,7 +8,7 @@ import pytest
 
 from vibesensor.domain import SpeedProfile
 from vibesensor.shared.boundaries.diagnostic_case import speed_profile_from_stats
-from vibesensor.use_cases.diagnostics.helpers import _speed_stats
+from vibesensor.use_cases.diagnostics.speed_profile_helpers import _speed_stats
 from vibesensor.use_cases.diagnostics.summary_builder import (
     PreparedRunData,
     RunAnalysis,
