+++ apps/server/vibesensor/use_cases/diagnostics/location_analysis.py
@@ -13,7 +13,7 @@ from vibesensor.shared.json_utils import as_float_or_none as _as_float
 from vibesensor.shared.json_utils import i18n_ref
 from vibesensor.shared.locations import has_any_wheel_location, is_wheel_location
 from vibesensor.shared.types.json_types import JsonObject
-from vibesensor.use_cases.diagnostics.helpers import _weighted_percentile
+from vibesensor.use_cases.diagnostics.math_utils import _weighted_percentile
 
 NEAR_TIE_DOMINANCE_THRESHOLD = 1.15
